requirements.txt
PySide6>=6.6.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0
chromadb>=0.5.0
duckduckgo-search>=6.0.0
wikipedia>=1.4.0
python-dotenv>=1.0.0
openai>=1.30.0
